Adding lazy load to images
I added a lazy load to the images and i also have add another custom style to tailwindcss.config.js for spacing 320px. I used that custom size for fix images. And i also added a library called react-responsive so i can check for a specific viewport and that's what i did to remove the height that was set on lazyload component. for now.